Refactor price logic to ol-utilities
