+++ frontends/ol-utilities/src/index.ts
@@ -6,6 +6,7 @@ export * from "./styles"
 
 export * from "./date/format"
 export * from "./learning-resources/learning-resources"
+export * from "./learning-resources/pricing"
 export * from "./strings/html"
 export * from "./predicates"
 export * from "./hooks"
